refactor(MFLP-82): Refactoring Auth & User applicationName.
