+++ buyhood-api-gateway/src/main/java/api/buyhood/gateway/BuyHoodApiGatewayApplication.java
@@ -7,10 +7,10 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-public class BuyhoodApiGatewayApplication {
+public class BuyHoodApiGatewayApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(BuyhoodApiGatewayApplication.class, args);
+		SpringApplication.run(BuyHoodApiGatewayApplication.class, args);
 	}
 
 }
