+++ buyhood-auth/src/test/java/api/buyhood/auth/BuyHoodAuthApplicationTests.java
@@ -1,10 +1,10 @@
-package api.buyhood.gateway;
+package api.buyhood.auth;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BuyhoodApiGatewayApplicationTests {
+class BuyHoodAuthApplicationTests {
 
 	@Test
 	void contextLoads() {
