+++ buyhood-eureka/src/test/java/api/buyhood/eureka/BuyHoodEurekaApplicationTests.java
@@ -1,10 +1,10 @@
-package api.buyhood.auth;
+package api.buyhood.eureka;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BuyhoodDomainAuthApplicationTests {
+class BuyHoodEurekaApplicationTests {
 
 	@Test
 	void contextLoads() {
